+++ examples/scriptengine/example.js
@@ -1,7 +1,3 @@
-var parser = new Packages.tv.ustream.yolo.module.parser.scriptengine.IScriptParser
-{
-    parse: function(line)
-    {
-        return {"line": line};
-    }
+function parse(line) {
+    return {"line": line.toUpperCase()};
 }